+++ packages/terra-doc-template/src/terra-dev-site/test/doc-template/NpmBadge.test.jsx
@@ -1,0 +1,19 @@
+import React from 'react';
+// eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
+import NpmBadge from '../../../NpmBadge';
+import { name } from '../../../../package.json';
+
+const packageUrl = 'https://engineering.cerner.com/terra-ui/components/terra-doc-template/doc-template/doc-template';
+const packageVersion = '2.13.0';
+
+const BadgeExamples = () => (
+  <div>
+    <br />
+    <NpmBadge packageName={name} />
+    <br />
+    <NpmBadge packageName={name} packageVersion={packageVersion} />
+    <NpmBadge packageName={name} packageUrl={packageUrl} packageVersion={packageVersion} />
+  </div>
+);
+
+export default BadgeExamples;